requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
langchain-core
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
transformers
